third_party/WebKit/LayoutTests/fast/filesystem/entry-points-missing-arguments.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>
description('Ensures that we get exceptions when we do not include required arguments for webkitRequestFileSystem and webkitResolveLocalFileSystemURL methods.');

if (window.webkitResolveLocalFileSystemURL) {
    shouldThrow('webkitResolveLocalFileSystemURL()', '"TypeError: Failed to execute \'webkitResolveLocalFileSystemURL\' on \'Window\': 2 arguments required, but only 0 present."');
    shouldThrow('webkitResolveLocalFileSystemURL("")', '"TypeError: Failed to execute \'webkitResolveLocalFileSystemURL\' on \'Window\': 2 arguments required, but only 1 present."');
    shouldThrow('webkitRequestFileSystem()', '"TypeError: Failed to execute \'webkitRequestFileSystem\' on \'Window\': 3 arguments required, but only 0 present."');
    shouldThrow('webkitRequestFileSystem(TEMPORARY)', '"TypeError: Failed to execute \'webkitRequestFileSystem\' on \'Window\': 3 arguments required, but only 1 present."');
    shouldThrow('webkitRequestFileSystem(TEMPORARY, 1)', '"TypeError: Failed to execute \'webkitRequestFileSystem\' on \'Window\': 3 arguments required, but only 2 present."');
} else
    debug("This test requires window.webkitResolveLocalFileSystemURL.");
</script>
